+++ flecsi/exec/fwd.hh
@@ -70,16 +70,19 @@ template<auto & Task,
   @tparam TASK          The user task.
     Its parameters must be copyable or a reference to a const, movable type.
     Any that is a pointer must be to a const type or to a function.
-    If \a ATTRIBUTES specifies an MPI task, parameters need merely be movable.
+    If \a ATTRIBUTES specifies an MPI task, parameters need merely be movable,
+    though the temporary objects to which any references bind must be movable
+    and must not themselves bind references to temporaries.
   @tparam ATTRIBUTES    The task attributes mask.
   @tparam ARGS The user-specified task arguments, implicitly converted to the
     parameter types for \a TASK.
     Certain FleCSI-defined parameter types accept particular, different
     argument types that serve as selectors for information stored by the
     backend; each type involved documents the correspondence.
-    Additionally, a parameter may be a \c std::vector of such a type or a \c
-    std::tuple that includes such a type; it accepts a \c std::vector of or a
-    \c std::tuple including the corresponding argument type.
+    If a parameter and its argument are each a (reference to a) \c std::vector
+    or a \c std::tuple (of the same size), their elements are treated as
+    separate parameters/arguments (with the unusual corollary that a
+    `std::vector<int>` matches a parameter of type `std::vector<long>`).
   \return a \ref future providing the value(s) returned from the task
 
   \ns.